libutil: move ev_zmq to libzmqutil
Problem: In an effort to consolidate zmq dependent code into
one library, ev_zmq in libutil is out of place.

Solution: Move ev_zmq from libutil to libzmqutil.  As a result,
libutil is no longer dependent on libzmq, and thus libflux-internal
is no longer dependent on libzmq.  Remove libzmq dependency
in libflux-internal.  Any binaries/libraries that need libzmq
and previously acquired it via libflux-internal, add libzmq dependency
directly.
