+++ src/common/libjob/Makefile.am
@@ -49,11 +49,13 @@ T_LOG_DRIVER = env AM_TAP_AWK='$(AWK)' $(SHELL) \
 test_ldadd = \
 	$(top_builddir)/src/common/libjob/libjob.la \
 	$(top_builddir)/src/common/libflux/libflux.la \
+	$(top_builddir)/src/common/libzmqutil/libzmqutil.la \
 	$(top_builddir)/src/common/libflux-internal.la \
 	$(top_builddir)/src/common/libtap/libtap.la \
 	$(JANSSON_LIBS) \
 	$(LIBPTHREAD) \
-	$(FLUX_SECURITY_LIBS)
+	$(FLUX_SECURITY_LIBS) \
+	$(ZMQ_LIBS)
 
 test_cppflags = \
 	$(AM_CPPFLAGS) \
