+++ src/common/libkvs/Makefile.am
@@ -60,10 +60,12 @@ T_LOG_DRIVER = env AM_TAP_AWK='$(AWK)' $(SHELL) \
 test_ldadd = \
 	$(top_builddir)/src/common/libkvs/libkvs.la \
 	$(top_builddir)/src/common/libflux/libflux.la \
+	$(top_builddir)/src/common/libzmqutil/libzmqutil.la \
 	$(top_builddir)/src/common/libflux-internal.la \
 	$(top_builddir)/src/common/libtap/libtap.la \
 	$(JANSSON_LIBS) \
-	$(LIBPTHREAD)
+	$(LIBPTHREAD) \
+	$(ZMQ_LIBS)
 
 test_cppflags = \
 	$(AM_CPPFLAGS) \
